Change the instrumentation process
